Create object to run compile-extensions scripts
- There are several places in the buildpack code that need to call
  compile extensions scripts, so we created a module for that purpose

[#126394949]
